backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
requests==2.32.3
dnspython==2.7.0
Flask-Limiter==3.10.1
supabase==2.27.3
twilio==9.10.0
